% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrases.R
\docType{data}
\name{phrases}
\alias{phrases}
\title{Bat pun phrases used in say()}
\format{
An object of class \code{character} of length 30.
}
\usage{
phrases
}
\description{
Bat pun phrases used in say()
}
\keyword{datasets}
