% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{say}
\alias{say}
\title{Bat Puns Told By a Computer Chiropteran}
\usage{
say(what)
}
\arguments{
\item{what}{Whatever text you want the bat to say. If not supplied, a random (amazing bat pun) phrase is chosen}
}
\value{
Bat with a pick-me-up message
}
\description{
Summon a bat for a cheerful message
}
\examples{
say("Hi, I'm Bat-thew McConaughey")
}
